Home

SparseMatricesCSR.jl

SparseMatricesCSR.SparseMatrixCSR

SparseMatricesCSR.SymSparseMatrixCSR

Base.count

Base.count

SparseArrays.findnz

SparseArrays.findnz

SparseArrays.nnz

SparseArrays.nnz

SparseArrays.nonzeros

SparseArrays.nonzeros

SparseArrays.nzrange

SparseArrays.nzrange

SparseMatricesCSR.finalize_coo!

SparseMatricesCSR.finalize_coo!

SparseMatricesCSR.finalize_coo!

SparseMatricesCSR.finalize_coo!

SparseMatricesCSR.push_coo!

SparseMatricesCSR.push_coo!

SparseMatricesCSR.push_coo!

SparseMatricesCSR.push_coo!

SparseMatricesCSR.sparsecsr

SparseMatricesCSR.symsparsecsr

SparseMatricesCSR.SparseMatrixCSR — Type.

struct SparseMatrixCSR{T,Ti<:Integer} <: AbstractSparseMatrix{T,Ti}

Matrix type for storing sparse matrices in the Compressed Sparse Row format. The standard way of constructing SparseMatrixCSR is through the sparsecsr function.

source

SparseMatricesCSR.SymSparseMatrixCSR — Type.

struct SymSparseMatrixCSR{T,Ti<:Integer} <: AbstractSparseMatrix{T,Ti}

Matrix type for storing symmetric sparse matrices in the Compressed Sparse Row format. The standard way of constructing SparseMatrixCSR is through the symsparsecsr function.

source

SparseMatricesCSR.finalize_coo! — Method.

function finalize_coo!(I,J,V,m,n)

Check and insert diagonal entries in COO vectors if needed.

source

SparseMatricesCSR.finalize_coo! — Method.

function finalize_coo!(::Type{SparseMatrixCSC},I,J,V,m,n)

Check and insert diagonal entries in COO vectors if needed.

source

SparseMatricesCSR.finalize_coo! — Method.

function finalize_coo!(::Type{SparseMatrixCSR},I,J,V,m,n)

Check and insert diagonal entries in COO vectors if needed.

source

SparseMatricesCSR.finalize_coo! — Method.

function finalize_coo!(::Type{SymSparseMatrixCSR},I,J,V,m,n)

Check and insert diagonal entries in COO vectors if needed.

source

SparseMatricesCSR.push_coo! — Method.

function push_coo!(I,J,V,ik,jk,vk)

Inserts entries in COO vectors for further building a SparseMatrixCSC.

source

SparseMatricesCSR.push_coo! — Method.

function push_coo!(::Type{SparseMatrixCSC},I,J,V,ik,jk,vk)

Inserts entries in COO vectors for further building a SparseMatrixCSC.

source

SparseMatricesCSR.push_coo! — Method.

function push_coo!(::Type{SparseMatrixCSR},I,J,V,ik,jk,vk)

Inserts entries in COO vectors for further building a SparseMatrixCSR.

source

SparseMatricesCSR.push_coo! — Method.

function push_coo!(::Type{SymSparseMatrixCSR},I,J,V,ik,jk,vk)

Inserts entries in COO vectors for further building a SymSparseMatrixCSR. It stores only the upper triangle, ignoring entries with (ik>jk) coordinates.

source

SparseMatricesCSR.sparsecsr — Method.

sparsecsr(I, J, V, [m, n, combine])

Create a sparse matrix S of dimensions m x n such that S[I[k], J[k]] = V[k]. The combine function is used to combine duplicates. If m and n are not specified, they are set to maximum(I) and maximum(J) respectively. If the combine function is not supplied, combine defaults to + unless the elements of V are Booleans in which case combine defaults to |. All elements of I must satisfy 1 <= I[k] <= m, and all elements of J must satisfy 1 <= J[k] <= n. Numerical zeros in (I, J, V) are retained as structural nonzeros; to drop numerical zeros, use dropzeros!.

source

SparseMatricesCSR.symsparsecsr — Method.

symsparsecsr(I, J, V, [m, n, combine])

Create a Symmetric sparse matrix S of dimensions m x n such that S[I[k], J[k]] = V[k], and m=n. The combine function is used to combine duplicates. If m and n are not specified, they are set to maximum(I) and maximum(J) respectively. If the combine function is not supplied, combine defaults to +. All elements of I must satisfy 1 <= I[k] <= m, and all elements of J must satisfy 1 <= J[k] <= n. Numerical zeros in (I, J, V) are retained as structural nonzeros; to drop numerical zeros, use dropzeros!.

source

Base.count — Method.

count(pred, S::SparseMatrixCSR) -> Integer

Count the number of elements in S for which predicate pred returns true.

source

Base.count — Method.

count(pred, S::SymSparseMatrixCSR) -> Integer

Count the number of elements in S for which predicate pred returns true.

source

SparseArrays.findnz — Method.

findnz(S::SparseMatrixCSR)

Return a tuple (I, J, V) where I and J are the row and column indices of the stored ("structurally non-zero") values in sparse matrix A, and V is a vector of the values.

source

SparseArrays.findnz — Method.

findnz(S::SymSparseMatrixCSR)

Return a tuple (I, J, V) where I and J are the row and column indices of the stored ("structurally non-zero") values in sparse matrix A, and V is a vector of the values.

source

SparseArrays.nnz — Method.

nnz(S::SparseMatrixCSR)

Returns the number of stored (filled) elements in a sparse array.

source

SparseArrays.nnz — Method.

nnz(S::SymSparseMatrixCSR)

Returns the number of stored (filled) elements in a sparse array.

source

SparseArrays.nonzeros — Method.

nonzeros(S::SparseMatrixCSR)

Return a vector of the structural nonzero values in sparse array S. This includes zeros that are explicitly stored in the sparse array. The returned vector points directly to the internal nonzero storage of A, and any modifications to the returned vector will mutate A as well.

source

SparseArrays.nonzeros — Method.

nonzeros(S::SymSparseMatrixCSR)

Return a vector of the structural nonzero values in sparse array S. This includes zeros that are explicitly stored in the sparse array. The returned vector points directly to the internal nonzero storage of A, and any modifications to the returned vector will mutate A as well.

source

SparseArrays.nzrange — Method.

nzrange(S::SparseMatrixCSR, row::Integer

Return the range of indices to the structural nonzero values of a sparse matrix row.

source

SparseArrays.nzrange — Method.

nzrange(S::SymSparseMatrixCSR, row::Integer)

Return the range of indices to the structural nonzero values of a sparse matrix row.

source